Set current directory when invoking compiler
While we pass the input and output parameters to the compiler as absolute paths,
under Windows `cl.exe` will generate intermediate files in the current
directory. We don't care about them, but they should be generated in the
temporary directory specific to the current rsconf instance so they can be
cleaned up on drop.