Don't panic on rustc_version() failure
We were originally panicking so that any errors or incorrect assumptions on our
end could be caught and fixed, but that's not really future-safe behavior and
surely we'll find out about the breakage some other way, even if not as soon.

We're already returning the version number wrapped an in an `Option` so we might
as well just return `None` in these cases.